% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{budgetToScheme}
\alias{budgetToScheme}
\title{budgetToScheme function}
\usage{
budgetToScheme(percentages, bsd)
}
\arguments{
\item{percentages}{Numeric vector with nStages+1 cells which are 
alloction for PIC and for each stage.}

\item{bsd}{List of breeding scheme data}
}
\value{
A revised bsd with the sizes of the stages fitting the percentages
}
\description{
Given breeding scheme budget percentages, return numbers of crosses and plots
The percentages are for PIC and for allocation to each VDP stage
Rules: 1. For PIC, nBreedingProg has to be at least minNBreedingProg
2. Each stage has to have fewer entries than the previous stage
If the percentages break the rules budgetToScheme returns a failure
The total budget allowed is given by bsd$budget minus location maintenance
}
\details{
Call this function after running specifyCosts.
}
\examples{
Assume stages of CET, PYT, UYT, so percentRanges needs 4 rows
The first stage needs more budget because of genotyping and development costs
percentages <- c(0.50, 0.30, 0.10, 0.10)
bsd <- budgetToScheme(bsd, percentages=percentages)

}
