% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{calcCurrentStatus}
\alias{calcCurrentStatus}
\title{calcCurrentStatus function}
\usage{
calcCurrentStatus(bsd)
}
\arguments{
\item{bsd}{List of breeding program data}
}
\value{
A vector extracting breeding population means and variances
}
\description{
Calculate mean and variance statistics on your current breeding program.
}
\details{
This function is only called internally by other functions used to specify the pipeline
}
