% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{calcMinMaxEntries}
\alias{calcMinMaxEntries}
\title{calcMinMaxEntries function}
\usage{
calcMinMaxEntries(bsd, budgetVec = NULL, includePIC = F)
}
\arguments{
\item{bsd}{List of breeding program data}

\item{budgetVec}{vector as comes out of calcBudget. If NULL look in bsd}

\item{includePIC}{logical should a min and max for nProg in PIC be figured?}
}
\value{
List with vectors for minimum entries and maximum entries and cost
per individual entry
}
\description{
calcMinMaxEntries function
}
\details{
Assume the location costs are fixed.
}
\examples{
minMaxList <- calcMinMaxEntries(bsd)

}
