% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{calcVDPcovMat}
\alias{calcVDPcovMat}
\title{calcVDPcovMat function}
\usage{
calcVDPcovMat(bsd)
}
\arguments{
\item{bsd}{List of breeding program data}
}
\value{
Covariance matrix of the true genotypic values of the variety
candidates with the phenotypic means from the VDP stages
}
\description{
calcVDPcovMat function
}
\details{
This function came essentially from the selectiongain package 
developed by Melchinger's group: Mi et al. 2014 Crop Science. 
Simplified just to the "LonginII" method.
}
\examples{
vdpCovMat <- calcVDPcovMat(bsd)

}
