% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFront.R
\name{findNonDom}
\alias{findNonDom}
\title{findNonDom function}
\usage{
findNonDom(mat, dir1Low = T, dir2Low = T, var1name = NULL, var2name = NULL)
}
\arguments{
\item{mat}{Real matrix with named columns}

\item{dir1Low}{Logical: desired direction in down for column 1}

\item{dir2Low}{Logical: desired direction in down for column 2}

\item{var1name}{Character: name of the column of column 1}

\item{var2name}{Character: name of the column of column 2}
}
\value{
nonDom Real matrix, with rows that are non-dominated with 
respect to var1 and var2 and sorted for var1
}
\description{
Find the non-dominated points in a matrix
}
\details{
Utility for Pareto front
}
