% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{loessPredCount}
\alias{loessPredCount}
\title{loessPredCount function}
\usage{
loessPredCount(
  resultMat,
  nSim = nrow(resultMat),
  xlim = NULL,
  ylim = NULL,
  budg1 = 1,
  budg2 = 2
)
}
\arguments{
\item{resultMat}{Tibble the simulation results}

\item{nSim}{Integer Predictions and counts for the first nSim simulations}

\item{xlim}{Real max and min for plot}

\item{ylim}{Real max and min for plot}

\item{budg1}{Integer which column of percent budget to make plot for x-axis}

\item{budg2}{Integer which column of percent budget to make plot for y-axis}
}
\value{
A list with bin counts and means and results from LOESS predictions
}
\description{
Summarize data from simulations in a form to make a nice hexbin plot
Return both counts in bins and bin means
}
\details{
Use the hexbin function to tabulate 2D bins in hexagonal array
}
