% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{makeCrosses}
\alias{makeCrosses}
\title{makeCrosses function}
\usage{
makeCrosses(bsd, optCont)
}
\arguments{
\item{bsd}{List of breeding program data}

\item{optCont}{Data.frame Optimal contributions coming from selectParents}
}
\value{
Updated bsd with new S0 progeny in breedingPop
}
\description{
Function to make crosses to get the next generation of breeding progeny
}
\details{
Takes the optimal contributions and generates random mating with
the correct number of progeny per parent. Self-fertilization is possible.
}
\examples{
bsd <- makeCrosses(bsd, optCont)]

}
