% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{makeCrossesBurnIn}
\alias{makeCrossesBurnIn}
\title{makeCrossesBurnIn function}
\usage{
makeCrossesBurnIn(bsd, parents)
}
\arguments{
\item{bsd}{List of breeding program data}

\item{parents}{String vector of ids of parents}
}
\value{
Updated bsd with new S0 progeny in breedingPop
}
\description{
Function to make crosses during the phenotypic selection burnin
}
\details{
Randomly mates the parents. Self-fertilization is possible.
}
\examples{
bsd <- makeCrossesBurnIn(bsd, parents)]

}
