% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{makeGrid}
\alias{makeGrid}
\title{makeGrid function}
\usage{
makeGrid(bsd)
}
\arguments{
\item{bsd}{List of breeding scheme data}
}
\value{
List with initial and end means and variances
}
\description{
Minimum and maximum budget percentages for each part and step sizes
These vectors are in bsd
Returns a list with all the outputs
}
\details{
Call this function to set up optimization
}
\examples{
bsd <- makeGrid(bsd)

}
