% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{makeNEntryGrid}
\alias{makeNEntryGrid}
\title{makeNEntryGrid function}
\usage{
makeNEntryGrid(
  nIndMin,
  nIndMax,
  step = NULL,
  minBudget = 0.9 * maxBudget,
  maxBudget,
  indCost
)
}
\arguments{
\item{nIndMin}{nStages vector minimum number of individuals at each stage}

\item{nIndMax}{nStages vector maximum number of individuals at each stage}

\item{step}{nStages vector: interval of number of individuals. If NULL will
increment for last stage and have the same number of steps for other stages}

\item{maxBudget}{the maximum cost of all evaluations}

\item{costProd}{per individual, how much does it cost to produce the seed?}

\item{costTest}{per individual, how much does it cost to run all the plots?}
}
\value{
List with feasible entry number schemes
}
\description{
makeNEntryGrid function
}
\details{
Call this function to set up optimization
}
\examples{
bsd <- makeNEntryGrid(bsd)

}
