% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{makeVarietyCandidates}
\alias{makeVarietyCandidates}
\title{makeVarietyCandidates function}
\usage{
makeVarietyCandidates(bsd, breedPopIDs = NULL)
}
\arguments{
\item{bsd}{List of breeding program data}

\item{breedPopIDs}{String vector of IDs of breeding individuals progenitors
of the variety candidates.  If NULL, the last nBreedingProg individuals used}
}
\value{
Updated bsd with new variety candidates in bsd$varietyCandidates
}
\description{
makeVarietyCandidates function
}
\details{
Here, creates DHs evenly distributed among the breeding progeny
in the last generation
}
\examples{
bsd <- makeVarietyCandidates(bsd)]

}
