% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{multistageGain}
\alias{multistageGain}
\title{multistageGain function}
\usage{
multistageGain(corr, truncPts, Vg = 1)
}
\arguments{
\item{corr}{matrix the correlation matrix from calcVDPcovMat}

\item{truncPts}{vector the truncation points that came from multistageTruncPt}

\item{Vg}{numeric the genotypic variance to provide a scale}
}
\value{
numeric the expected gain from selection
}
\description{
multistageGain function
}
\details{
This function came essentially from the selectiongain package 
but then I simplified it very much by just using a current R package
}
\examples{
vdpCovMat <- multistageGain(corr, truncPts)

}
