% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{multistageTruncPt}
\alias{multistageTruncPt}
\title{multistageTrucPt function}
\usage{
multistageTruncPt(alpha, corr = NA)
}
\arguments{
\item{alpha}{vector the selected fractions from one VDP stage to the next}

\item{corr}{matrix the correlation matrix from calcVDPcovMat}

\item{alg}{function the truncation point calculation algorithm from
package mvtnorm. Options are Miwa() and GenzBretz(). The latter is slower}
}
\value{
Covariance matrix of the true genotypic values of the variety
candidates with the phenotypic means from the VDP stages
}
\description{
multistageTrucPt function
}
\details{
This function came essentially from the selectiongain package 
developed by Melchinger's group: Mi et al. 2014 Crop Science. 
Simplified just to the "LonginII" method.
}
\examples{
vdpCovMat <- calcVDPcovMat(bsd)

}
