% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{plotLoessPred}
\alias{plotLoessPred}
\title{plotLoessPred function}
\usage{
plotLoessPred(
  nSim = nrow(resultMat),
  resultMat,
  xlim = NULL,
  ylim = NULL,
  budg1 = 1,
  budg2 = 2,
  binMeanContrast = 3
)
}
\arguments{
\item{nSim}{Integer Predictions and counts for the first nSim simulations}

\item{resultMat}{Tibble the simulation results}

\item{xlim}{Real max and min for plot}

\item{ylim}{Real max and min for plot}

\item{budg1}{Integer which column of percent budget to make plot for x-axis}

\item{budg2}{Integer which column of percent budget to make plot for y-axis}

\item{binMeanContrast}{Numeric a higher value makes the peak stand out more}
}
\value{
A list with bin counts and means and results from LOESS predictions
}
\description{
Make a nice hexbin plot with the summary from loessPredCount
}
\details{
Makes a plot
}
