% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{runBatch}
\alias{runBatch}
\title{runBatch function}
\usage{
runBatch(batchBudg, bsd)
}
\arguments{
\item{batchBudg}{List of budget percentage vectors for 
the population improvement cycle and the variety development pipeline}

\item{bsd}{List of breeding scheme data}
}
\value{
Tibble with initial and end means and variances
}
\description{
Run a batch of replications with specified budget percentages
These percentages have to be valid in terms of following rules
}
\details{
Call this function to test out a batch of different budgets
}
\examples{
bsd <- runBatch(batchBudg, bsd)

}
