% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{selectParentsBurnIn}
\alias{selectParentsBurnIn}
\title{selectParentsBurnIn function}
\usage{
selectParentsBurnIn(bsd)
}
\arguments{
\item{bsd}{The breeding scheme data list}
}
\value{
Vector of entry IDs
}
\description{
Function to select varieties to cross during phenotypic selection burnin
phase, from variety candidates, by truncation selection
}
\details{
Accesses data in phenoRecords to pick the highest among candidates
}
\examples{
entries <- selectParentsBurnIn(bsd)]

}
