% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{calcBudget}
\alias{calcBudget}
\title{calcBudget function}
\usage{
calcBudget(bsd)
}
\arguments{
\item{bsd}{List of breeding scheme data}
}
\value{
The bsd list with updated intermediate and total costs
}
\description{
Once the costs are specified in bsd, this function calculates the total annual budget for the breeding scheme
}
\details{
Call this function once costs have been specified
}
\examples{
bsd <- calcBudget(bsd)

}
