% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{chooseTrialEntries}
\alias{chooseTrialEntries}
\title{chooseTrialEntries function}
\usage{
chooseTrialEntries(bsd, toTrial, fromTrial = NULL)
}
\arguments{
\item{bsd}{The breeding scheme data list}

\item{toTrial}{String the trial type the entries are going to}

\item{fromTrial}{String the trial type the entries are coming from}
}
\value{
bsd updated with a vector of entry IDs in bsd$entries
}
\description{
Function to select which varieties to advance to the next stage assuming
they are IID (i.e., estimating genotypic values), by truncation selection
}
\details{
Accesses all data in phenoRecords to pick the highest 
performing candidates.
}
\examples{
bsd <- chooseTrialEntries(bsd, toTrial, fromTrial)]

}
