% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{findRedoBudgets}
\alias{findRedoBudgets}
\title{findRedoBudgets function}
\usage{
findRedoBudgets(bsd, results)
}
\arguments{
\item{bsd}{List of breeding scheme data}

\item{results}{List with the results of simulations so far}
}
\value{
List of two vectors one with high gain budgets
and one with uncertainty budgets
}
\description{
Based on simulations run so far, find the budgets that gave the highest
gain and the budgets with high gain but also high uncertainty
}
\details{
Call this function to choose which budgets to try again
}
\examples{
bsd <- findRedoBudgets(bsd, results)

}
