% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostOptimization.R
\name{runWithBudget}
\alias{runWithBudget}
\title{runWithBudget function}
\usage{
runWithBudget(percentages, bsd, returnBSD = F)
}
\arguments{
\item{percentages}{Double vector 1 + nStages with budget percentages for 
the population improvement cycle and the variety development pipeline}

\item{bsd}{List of breeding scheme data}

\item{returnBSD}{Logical if you want to return bsd because this is a one-off
rather than being called repeatedly for optimization}
}
\value{
Vector with initial and end means and variances
}
\description{
Run one replication of the scheme with specified budget percentages
These percentages have to be valid in terms of following rules
}
\details{
Call this function to test out the response of different budgets
}
\examples{
bsd <- runWithBudget(percentages, bsd)

}
